# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Moments API",
        "contact": {},
        "version": "0.2.1"
    },
    "paths": {
...
        "/api/memo/get": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memo"
                ],
                "summary": "获取单个memo详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "memoID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.Memo"
                        }
                    }
                }
            }
        },
        "/api/memo/getDoubanBookInfo": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memo"
                ],
                "summary": "获取豆瓣读书详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "豆瓣读书ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录TOKEN",
                        "name": "x-api-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.DoubanBook"
                        }
                    }
                }
            }
        },
        "/api/memo/getDoubanMovieInfo": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memo"
                ],
                "summary": "获取豆瓣电影详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "豆瓣电影ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录TOKEN",
                        "name": "x-api-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.DoubanMovie"
                        }
                    }
                }
            }
        },
        "/api/memo/getFaviconAndTitle": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memo"
                ],
                "summary": "获取外部链接的相关信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "memoID",
                        "name": "url",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录TOKEN",
                        "name": "x-api-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.externalWebsite"
                        }
                    }
                }
            }
        },
        "/api/memo/like": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memo"
                ],
                "summary": "点赞memo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "memoID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/memo/list": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memo"
                ],
                "summary": "查询memos列表",
                "parameters": [
                    {
                        "description": "查询memos列表",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.ListMemoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.memoListResp"
                        }
                    }
                }
            }
        },
        "/api/memo/remove": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memo"
                ],
                "summary": "删除memos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "memoID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录TOKEN",
                        "name": "x-api-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/memo/save": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memo"
                ],
                "summary": "保存memos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录TOKEN",
                        "name": "x-api-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "保存memos",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.SaveMemoReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/memo/setPinned": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memo"
                ],
                "summary": "置顶单个memo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "memoID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "登录TOKEN",
                        "name": "x-api-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.Memo": {
            "type": "object",
            "properties": {
                "bilibiliUrl": {
                    "type": "string"
                },
                "commentCount": {
                    "type": "integer"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Comment"
                    }
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "ext": {
                    "type": "string"
                },
                "externalFavicon": {
                    "type": "string"
                },
                "externalTitle": {
                    "type": "string"
                },
                "externalUrl": {
                    "type": "string"
                },
                "favCount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "imgs": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "music163Url": {
                    "type": "string"
                },
                "pinned": {
                    "type": "boolean"
                },
                "showType": {
                    "type": "integer"
                },
                "tags": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/db.User"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "handler.externalWebsite": {
            "type": "object",
            "properties": {
                "favicon": {
                    "description": "favicon",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                }
            }
        },
...
    }
}